requirements.txt
prometheus_client==0.18.0
python-dotenv==1.0.0
web3==6.11.3
uniswap-python==0.7.1
